Update README.md
This is a cheap ($30-$40) Bluetooth thermal printer. It's greatest
weakness is that it may have to be used only the the android/iphone app. Other than that, it offers a "steward"(partner/promoter) a way to onramp a group of people into Web3 without them having to install an App. The print-out acts as their entire wallet. A large group of people (10-100 or more) can thus be given unique crypto-bearing fast wallets/paper wallets. They can migrate their private key into a better tool later.